requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.5
Flask-ReCaptcha==0.4.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
py-bcrypt==0.4
PyYAML==6.0
requests==2.31.0
urllib3==2.0.3
Werkzeug==2.2.3
